# docs/assets/api.yml
openapi: 3.0.0
info:
  title: Robinson API
  version: 0.1.0
  description: Robinson RESTful API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
  contact:
    name: Ryan McCartney
    url: https://ryan.mccartney.info/robinson
    email: hello@mccartney.info
servers:
  - url: http://localhost:80/api/
paths:
  /books:
    get:
      summary: Get a list of all books
      tags:
        - books
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    post:
      summary: Add a new book
      tags:
        - books
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  /books/orphaned:
    get:
      summary: Get a list of all books that don't have a shelf (Orphaned)
      tags:
        - books
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  /books/favourites:
    get:
      summary: Get a list of all books that have been favourite'd
      tags:
        - books
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  /books/progress:
    get:
      summary: Get a list of all books that have been started (have progress)
      tags:
        - books
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  /books/new:
    get:
      summary: Get a list of all books in order of newest
      tags:
        - books
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                records:
                  type: number
                  description: Number of records to return
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/books/case/{caseId}":
    get:
      summary: Get a list of all books in a case
      tags:
        - books
      parameters:
        - in: path
          name: caseId
          schema:
            type: string
          required: true
          description: The case ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/books/shelf/{shelfId}":
    get:
      summary: Get a list of all books on a shelf
      tags:
        - books
      parameters:
        - in: path
          name: shelfId
          schema:
            type: string
          required: true
          description: The shelf ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/books/cover/{bookId}":
    get:
      summary: Get a book cover by it's ID
      tags:
        - books
      parameters:
        - in: path
          name: bookId
          schema:
            type: string
          required: true
          description: The book ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/books/{bookId}":
    get:
      summary: Get a book by it's ID
      tags:
        - books
      parameters:
        - in: path
          name: bookId
          schema:
            type: string
          required: true
          description: The book ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    put:
      summary: Update a book by it's ID
      tags:
        - books
      parameters:
        - in: path
          name: bookId
          schema:
            type: string
          required: true
          description: The book ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
    delete:
      summary: Delete a book by it's ID
      tags:
        - books
      parameters:
        - in: path
          name: bookId
          schema:
            type: string
          required: true
          description: The book ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
  "/books/ebook/{bookId}":
    post:
      summary: Upload an ebook
      tags:
        - books
      parameters:
        - in: path
          name: bookId
          schema:
            type: string
          required: true
          description: The book ID string
      responses:
        "200":
          description: Success
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "405":
          description: Incorrect request data
        "500":
          description: Error
...